set a route
