+++ app/controllers/users_controller.rb
@@ -52,14 +52,6 @@ class UsersController < ApplicationController
   end
 
   # Before filters
-  # Confirms a logged-in user.
-  def logged_in_user
-    unless logged_in?
-      store_location
-      flash[:danger] = "Please log in."
-      redirect_to login_url
-    end
-  end
 
   # confirms the correct user.
   def correct_user
