+++ config/routes.rb
@@ -1,11 +1,9 @@
 Rails.application.routes.draw do
-  get 'password_resets/new'
-
-  get 'password_resets/edit'
 
   resources :users
   resources :account_activations, only: [:edit]
   resources :password_resets, only: [:new, :create, :edit, :update]
+  resources :microposts, only: [:create, :destroy]
 
   root 'static_pages#home'
   get '/help',      to: 'static_pages#help'
@@ -15,5 +13,7 @@ Rails.application.routes.draw do
   get '/login',     to: 'sessions#new'
   post '/login',    to: 'sessions#create'
   delete '/logout', to: 'sessions#destroy'
+  get 'password_resets/edit'
+  get 'password_resets/new'
 
 end
