+++ test/fixtures/microposts.yml
@@ -19,6 +19,7 @@ most_recent:
   user: michael
 
 <% 30.times do |n| %>
+micropost_<%= n %>:
     content: <%= Faker::Lorem.sentence(5) %>
     created_at: <%= 42.days.ago %>
     user: michael
